requirements.txt

build
pytest
requests
twine
cryptography

cloud_foundry
PyJWT
